+++ src/constants/index.js
@@ -1,0 +1,2 @@
+const ErrorCodeRecordNotExist='P2025';
+module.exports={ ErrorCodeRecordNotExist};
