+++ src/routes/api/index.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
 router.use('/po-notes', require('./poNotes.routes'));
-router.use('/dsm/team-requests',require('./dsm/teamRequests.routes'))
+router.use('/dsm/team-requests',require('./dsm/teamRequests.routes'));
 router.use('/dsm/announcements',require('./dsm/announcements.routes'));
 module.exports = router;
